+++ src/utils.js
@@ -1,8 +1,22 @@
+import dayjs from "dayjs";
+import relativeTime from "dayjs/plugin/relativeTime";
+dayjs.extend(relativeTime);
+
 export const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
 
   return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+export const generateRecordDay = () => {
+  const daysAgo = getRandomInteger(0, 14);
+
+  if (daysAgo <= 7 && !!daysAgo) {
+    return dayjs().subtract(daysAgo, `day`).fromNow();
+  }
+
+  return daysAgo ? dayjs().subtract(daysAgo, `day`).format(`YYYY/MM/DD HH:mm`) : `today`;
 };
 
 export const RenderPosition = {
@@ -47,3 +61,4 @@ export const createElement = (template) => {
 // Единственный нюанс, что HTML в строке должен иметь общую обёртку,
 // то есть быть чем-то вроде <nav><a>Link 1</a><a>Link 2</a></nav>,
 // а не просто <a>Link 1</a><a>Link 2</a>
+
